+++ old/ui/UserInterface.java
@@ -1,11 +1,10 @@
-package net.fexcraft.app.fmt.ui_old;
+package net.fexcraft.app.fmt.ui;
 
 import java.util.ArrayList;
 
 import org.lwjgl.input.Mouse;
 import org.lwjgl.opengl.GL11;
 
-import net.fexcraft.app.fmt.FMTB;
 import net.fexcraft.app.fmt.FMTGLProcess;
 import net.fexcraft.app.fmt.ui_old.general.ControlsAdjuster;
 import net.fexcraft.app.fmt.ui_old.general.DialogBox;
@@ -15,18 +14,16 @@ import net.fexcraft.app.fmt.ui_old.general.SettingsBox;
 import net.fexcraft.app.fmt.ui_old.general.TextField;
 import net.fexcraft.app.fmt.ui_old.general.Toolbar;
 import net.fexcraft.app.fmt.utils.RayCoastAway;
-import net.fexcraft.app.fmt.utils.SessionHandler;
 import net.fexcraft.app.fmt.utils.Settings;
-import net.fexcraft.lib.common.Static;
-import net.fexcraft.lib.common.math.Time;
 
 /**
  * @author Ferdinand Calo' (FEX___96)
  *
  */
-public class UserInterface {
+public class UserInterface extends AbstractElement {
 
-	public static float scale_x, scale_y;//, scale;
+	public static float scale_x, scale_y, divscal;
+	public static int scale;
 	public static Element SELECTED = null;
 	public static Toolbar TOOLBAR;
 	public static DialogBox DIALOGBOX;
@@ -36,27 +33,27 @@ public class UserInterface {
 	//
 	private ArrayList<Element> elements = new ArrayList<>();
 	private FMTGLProcess root;
+	private float[] clearcolor;
 
 	public UserInterface(FMTGLProcess main){
-		this.root = main; /*root.setupUI(this);*/ rescale();
+		super(null, "root"); this.root = main; root.setupUI(this); this.rescale();
+		this.clearcolor = Settings.getBackGroundColor();
+		//
+		elements.add(new Element(this, "test0").setPosition(10, 10).setSize(100, 20));
 	}
 	
 	public void rescale(){
+		for(Element elm : elements) elm.rescale();
 		scale_x = root.getDisplayMode().getWidth();
 		scale_y = root.getDisplayMode().getHeight();
-		int facts = 1, uis = Settings.ui_scale(); if(uis < 0) uis = 1000;
-        while(facts < uis && scale_x / (facts + 1) >= 320 && scale_y / (facts + 1) >= 240) facts++;
-        scale_x = scale_x / facts; scale_y = scale_y / facts;
+		scale = 1; int uis = Settings.ui_scale(); if(uis < 0) uis = 1000;
+        while(scale < uis && scale_x / (scale + 1) >= 320 && scale_y / (scale + 1) >= 240) scale++;
+        scale_x = scale_x / scale; scale_y = scale_y / scale;
         scale_x = (float)Math.ceil(scale_x); scale_y = (float)Math.ceil(scale_y);
-        //scale = Math.min(scale_x, scale_y);
-		width = (int)scale_x; height = (int)scale_y;
+		width = (int)scale_x; height = (int)scale_y; divscal = 1f / scale;
 	}
-	
-	private int width, height;
-	private float[] clearcolor;
 
-	public void render(boolean bool){
-		//width = root.getDisplayMode().getWidth(); height = root.getDisplayMode().getHeight();
+	public void render(boolean screenshot){
 		{
 			GL11.glPushMatrix();
 	        GL11.glMatrixMode(GL11.GL_PROJECTION);
@@ -70,11 +67,10 @@ public class UserInterface {
 		//
 		GL11.glLoadIdentity();
 		GL11.glDepthFunc(GL11.GL_ALWAYS);
-		if(bool){
-			tmelm.render(width, height); logintxt.render(width, height);
+		if(screenshot){
+			//tmelm.render(width, height); logintxt.render(width, height);
 		}
 		else{
-			//for(OldElement elm : oldelements) elm.render(width, height);
 			for(Element elm : elements) elm.render(width, height);
 		}
 		GL11.glDepthFunc(GL11.GL_LESS);
@@ -85,15 +81,13 @@ public class UserInterface {
 	        GL11.glMatrixMode(GL11.GL_MODELVIEW);
 	        GL11.glPopMatrix();
 	        GL11.glDepthFunc(GL11.GL_LEQUAL);
-	        //GL11.glClearColor(0.5f, 0.5f, 0.5f, 0.2f);
-	    	if(clearcolor == null){ clearcolor = Settings.getBackGroundColor(); }
 	    	GL11.glClearColor(clearcolor[0], clearcolor[1], clearcolor[2], clearcolor[3]);
 	        GL11.glClearDepth(1.0);
 	        GL11.glPopMatrix();
 		}
 	}
 	
-	private Element tmelm = new TextField(null, "text", 4, 4, 500){
+	/*private Element tmelm = new TextField(null, "text", 4, 4, 500){
 		@Override
 		public void renderSelf(int rw, int rh){
 			this.y = rh - FMTB.get().getDisplayMode().getHeight() + 4;
@@ -131,7 +125,7 @@ public class UserInterface {
 			}
 			super.renderSelf(rw, rh);
 		}
-	};
+	};*/
 
 	public boolean isAnyHovered(){
 		boolean bool = false;
@@ -178,5 +172,10 @@ public class UserInterface {
 	}
 	
 	public ArrayList<Element> getElements(){ return elements; }
+
+	@Override
+	public AbstractElement getRoot(){
+		return null;
+	}
 	
 }
