+++ src/net/fexcraft/app/fmt/FMTB.java
@@ -30,23 +30,12 @@ import net.arikia.dev.drpc.DiscordEventHandlers;
 import net.arikia.dev.drpc.DiscordRPC;
 import net.fexcraft.app.fmt.demo.ModelT1P;
 import net.fexcraft.app.fmt.porters.PorterManager;
-import net.fexcraft.app.fmt.ui_old.Dialog;
-import net.fexcraft.app.fmt.ui_old.FontRenderer;
-import net.fexcraft.app.fmt.ui_old.UserInterface;
-import net.fexcraft.app.fmt.ui_old.editor.Editor;
-import net.fexcraft.app.fmt.ui_old.editor.GeneralEditor;
-import net.fexcraft.app.fmt.ui_old.editor.ModelGroupEditor;
-import net.fexcraft.app.fmt.ui_old.editor.PreviewEditor;
-import net.fexcraft.app.fmt.ui_old.editor.TextureEditor;
-import net.fexcraft.app.fmt.ui_old.general.ControlsAdjuster;
-import net.fexcraft.app.fmt.ui_old.general.Crossbar;
-import net.fexcraft.app.fmt.ui_old.general.DialogBox;
-import net.fexcraft.app.fmt.ui_old.general.NFC;
-import net.fexcraft.app.fmt.ui_old.general.SettingsBox;
-import net.fexcraft.app.fmt.ui_old.general.TextField;
-import net.fexcraft.app.fmt.ui_old.general.Toolbar;
-import net.fexcraft.app.fmt.ui_old.tree.HelperTree;
-import net.fexcraft.app.fmt.ui_old.tree.ModelTree;
+import net.fexcraft.app.fmt.ui.Dialog;
+import net.fexcraft.app.fmt.ui.FontRenderer;
+import net.fexcraft.app.fmt.ui.UserInterface;
+import net.fexcraft.app.fmt.ui.editor.Editor;
+import net.fexcraft.app.fmt.ui.general.DialogBox;
+import net.fexcraft.app.fmt.ui.general.TextField;
 import net.fexcraft.app.fmt.utils.Backups;
 import net.fexcraft.app.fmt.utils.DiscordUtil;
 import net.fexcraft.app.fmt.utils.GGR;
@@ -324,7 +313,7 @@ public class FMTB implements FMTGLProcess {
 		TextureManager.loadTexture("icons/editors/minimized", null);
 		TextureManager.loadTexture("icons/editors/expanded", null);
 		//
-		ui.getElements().add(UserInterface.DIALOGBOX = new DialogBox());
+		/*ui.getElements().add(UserInterface.DIALOGBOX = new DialogBox());
 		ui.getElements().add(UserInterface.SETTINGSBOX = new SettingsBox());
 		ui.getElements().add(UserInterface.FILECHOOSER = new NFC());
 		ui.getElements().add(UserInterface.CONTROLS = new ControlsAdjuster());
@@ -337,7 +326,7 @@ public class FMTB implements FMTGLProcess {
 		ui.getElements().add(new PreviewEditor());
 		//render last
 		ui.getElements().add(UserInterface.TOOLBAR = new Toolbar());
-		ui.getElements().add(new Crossbar());
+		ui.getElements().add(new Crossbar());*/
 		FMTB.MODEL.updateFields();
 	}
