+++ src/net/fexcraft/app/fmt/ui/Dialog.java
@@ -1,4 +1,4 @@
-package net.fexcraft.app.fmt.ui_old;
+package net.fexcraft.app.fmt.ui;
 
 import java.util.ArrayList;
