+++ src/net/fexcraft/app/fmt/ui/FontRenderer.java
@@ -1,4 +1,4 @@
-package net.fexcraft.app.fmt.ui_old;
+package net.fexcraft.app.fmt.ui;
 
 import java.awt.Color;
 import java.awt.Font;
