+++ src/net/fexcraft/app/fmt/ui/general/Crossbar.java
@@ -1,7 +1,7 @@
-package net.fexcraft.app.fmt.ui_old.general;
+package net.fexcraft.app.fmt.ui.general;
 
-import net.fexcraft.app.fmt.ui_old.Dialog;
-import net.fexcraft.app.fmt.ui_old.Element;
+import net.fexcraft.app.fmt.ui.Dialog;
+import net.fexcraft.app.fmt.ui.Element;
 import net.fexcraft.app.fmt.utils.TextureManager;
 
 /**
