+++ src/net/fexcraft/app/fmt/ui/general/DialogBox.java
@@ -1,16 +1,16 @@
 /**
  * 
  */
-package net.fexcraft.app.fmt.ui_old.general;
+package net.fexcraft.app.fmt.ui.general;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui_old.Dialog;
-import net.fexcraft.app.fmt.ui_old.Element;
-import net.fexcraft.app.fmt.ui_old.FontRenderer;
-import net.fexcraft.app.fmt.ui_old.UserInterface;
+import net.fexcraft.app.fmt.ui.Dialog;
+import net.fexcraft.app.fmt.ui.Element;
+import net.fexcraft.app.fmt.ui.FontRenderer;
+import net.fexcraft.app.fmt.ui.UserInterface;
 import net.fexcraft.app.fmt.utils.TextureManager;
 import net.fexcraft.lib.common.math.RGB;
