+++ src/net/fexcraft/app/fmt/ui/general/HoverMenu.java
@@ -1,8 +1,8 @@
-package net.fexcraft.app.fmt.ui_old.general;
+package net.fexcraft.app.fmt.ui.general;
 
 import java.util.ArrayList;
 
-import net.fexcraft.app.fmt.ui_old.Element;
+import net.fexcraft.app.fmt.ui.Element;
 
 /**
  *
