+++ src/net/fexcraft/app/fmt/ui/tree/ModelTree.java
@@ -1,13 +1,13 @@
-package net.fexcraft.app.fmt.ui_old.tree;
+package net.fexcraft.app.fmt.ui.tree;
 
 import org.lwjgl.input.Mouse;
 import org.lwjgl.opengl.GL11;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui_old.FontRenderer;
-import net.fexcraft.app.fmt.ui_old.UserInterface;
-import net.fexcraft.app.fmt.ui_old.editor.Editor;
-import net.fexcraft.app.fmt.ui_old.general.DialogBox;
+import net.fexcraft.app.fmt.ui.FontRenderer;
+import net.fexcraft.app.fmt.ui.UserInterface;
+import net.fexcraft.app.fmt.ui.editor.Editor;
+import net.fexcraft.app.fmt.ui.general.DialogBox;
 import net.fexcraft.app.fmt.utils.GGR;
 import net.fexcraft.app.fmt.utils.Settings;
 import net.fexcraft.app.fmt.wrappers.PolygonWrapper;
