+++ src/net/fexcraft/app/fmt/utils/SessionHandler.java
@@ -4,7 +4,7 @@ import java.io.File;
 import com.google.gson.JsonObject;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui_old.general.DialogBox;
+import net.fexcraft.app.fmt.ui.general.DialogBox;
 import net.fexcraft.lib.common.json.JsonUtil;
 import net.fexcraft.lib.common.utils.HttpUtil;
 import net.fexcraft.lib.common.utils.Print;
