+++ src/net/fexcraft/app/fmt/wrappers/GroupCompound.java
@@ -11,11 +11,11 @@ import java.util.stream.Collectors;
 import org.lwjgl.opengl.GL11;
 
 import net.fexcraft.app.fmt.FMTB;
-import net.fexcraft.app.fmt.ui_old.editor.ContainerButton;
-import net.fexcraft.app.fmt.ui_old.editor.Editor;
-import net.fexcraft.app.fmt.ui_old.editor.TextureEditor;
-import net.fexcraft.app.fmt.ui_old.general.DialogBox;
-import net.fexcraft.app.fmt.ui_old.general.TextField;
+import net.fexcraft.app.fmt.ui.editor.ContainerButton;
+import net.fexcraft.app.fmt.ui.editor.Editor;
+import net.fexcraft.app.fmt.ui.editor.TextureEditor;
+import net.fexcraft.app.fmt.ui.general.DialogBox;
+import net.fexcraft.app.fmt.ui.general.TextField;
 import net.fexcraft.app.fmt.utils.RayCoastAway;
 import net.fexcraft.app.fmt.utils.TextureManager;
 import net.fexcraft.app.fmt.utils.TextureManager.Texture;
