+++ src/net/fexcraft/app/fmt/wrappers/PolygonWrapper.java
@@ -7,7 +7,7 @@ import org.lwjgl.opengl.GL11;
 
 import com.google.gson.JsonObject;
 
-import net.fexcraft.app.fmt.ui_old.editor.TextureEditor;
+import net.fexcraft.app.fmt.ui.editor.TextureEditor;
 import net.fexcraft.app.fmt.utils.Settings;
 import net.fexcraft.lib.common.math.RGB;
 import net.fexcraft.lib.common.math.Vec3f;
